requirements.txt
streamlit
pdfplumber
sentence-transformers
scikit-learn
google-generativeai
python-dotenv
